refactor: cleanup update activities function
